digital-wardrobe/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-jwt
Pillow
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap3==11.0.0
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
whitenoise==4.1.2
